Split logic into two files: app.js and server.js
