+++ package.json
@@ -2,7 +2,7 @@
   "name": "pro04_back_jason_tanya_riley",
   "version": "1.0.0",
   "description": "> Premier purveyors of playful paraphenalia",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
