+++ server.js
@@ -1,0 +1,9 @@
+// server.js
+const app = require('./app');
+
+const port = process.env.PORT || 3000;
+
+// Start the server
+app.listen(port, () => {
+  console.log(`Server running at http://localhost:${port}`);
+});
